refactor: reorganized the codebase Reorganized files
