+++ server/api/bans.js
@@ -1,11 +1,11 @@
 import { Router } from 'express';
 
 import dispatcher from '../helpers/dispatcher.js';
-import globalUtils from '../helpers/globalutils.js';
-import { logText } from '../helpers/logger.ts';
+import globalUtils from '../helpers/utils/globalutils.js';
+import { logText } from '../helpers/utils/logger.ts';
 import { guildPermissionsMiddleware, rateLimitMiddleware } from '../helpers/middlewares.js';
 const router = Router({ mergeParams: true });
-import errors from '../helpers/errors.js';
+import errors from '../helpers/consts/errors.js';
 import lazyRequest from '../helpers/lazyRequest.js';
 import quickcache from '../helpers/quickcache.js';
 import Watchdog from '../helpers/watchdog.js';
