+++ server/api/entitlements.js
@@ -1,9 +1,9 @@
 import { Router } from 'express';
 
-import { logText } from '../helpers/logger.ts';
-import { generate } from '../helpers/snowflake.js';
+import { logText } from '../helpers/utils/logger.ts';
+import { generate } from '../helpers/utils/snowflake.js';
 const router = Router({ mergeParams: true });
-import { response_500 } from '../helpers/errors.js';
+import { response_500 } from '../helpers/consts/errors.js';
 import { cacheFor } from '../helpers/quickcache.js';
 
 router.param('code', async (req, _, next, code) => {
