+++ server/api/integrations.js
@@ -1,10 +1,10 @@
 import { Router } from 'express';
 
-import { logText } from '../helpers/logger.ts';
+import { logText } from '../helpers/utils/logger.ts';
 import { rateLimitMiddleware } from '../helpers/middlewares.js';
 import { cacheFor } from '../helpers/quickcache.js';
 const router = Router({ mergeParams: true });
-import { response_500 } from '../helpers/errors.js';
+import { response_500 } from '../helpers/consts/errors.js';
 import { middleware } from '../helpers/watchdog.js';
 
 router.get(
