+++ server/api/oauth2/applications.js
@@ -1,8 +1,8 @@
 import { Router } from 'express';
 
-import { logText } from '../../helpers/logger.ts';
+import { logText } from '../../helpers/utils/logger.ts';
 const router = Router({ mergeParams: true });
-import errors from '../../helpers/errors.js';
+import errors from '../../helpers/consts/errors.js';
 
 router.param('applicationid', async (req, res, next, applicationid) => {
   req.application = await global.database.getApplicationById(applicationid);
