+++ server/api/pins.js
@@ -1,8 +1,8 @@
 import { Router } from 'express';
 
 import dispatcher from '../helpers/dispatcher.js';
-import errors from '../helpers/errors.js';
-import { logText } from '../helpers/logger.ts';
+import errors from '../helpers/consts/errors.js';
+import { logText } from '../helpers/utils/logger.ts';
 import { channelMiddleware } from '../helpers/middlewares.js';
 import quickcache from '../helpers/quickcache.js';
