+++ server/api/tutorial.js
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 
-import { response_500 } from '../helpers/errors.js';
-import { logText } from '../helpers/logger.ts';
+import { response_500 } from '../helpers/consts/errors.js';
+import { logText } from '../helpers/utils/logger.ts';
 const router = Router();
 
 router.get('/', async (req, res) => {
