+++ server/api/users/me/billing.js
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 const router = Router();
-import Snowflake from '../../../helpers/snowflake.js';
+import Snowflake from '../../../helpers/utils/snowflake.js';
 
 router.param('guildid', async (req, _, next, guildid) => {
   req.guild = await global.database.getGuildById(guildid);
