+++ server/api/users/me/index.js
@@ -1,11 +1,11 @@
 import { Router } from 'express';
 
-import globalUtils from '../../../helpers/globalutils.js';
-import { logText } from '../../../helpers/logger.ts';
+import globalUtils from '../../../helpers/utils/globalutils.js';
+import { logText } from '../../../helpers/utils/logger.ts';
 import { guildMiddleware, rateLimitMiddleware } from '../../../helpers/middlewares.js';
 const router = Router();
 import dispatcher from '../../../helpers/dispatcher.js';
-import errors from '../../../helpers/errors.js';
+import errors from '../../../helpers/consts/errors.js';
 import quickcache from '../../../helpers/quickcache.js';
 import Watchdog from '../../../helpers/watchdog.js';
 import relationships from '../relationships.js';
