+++ server/handlers/gateway.js
@@ -1,5 +1,5 @@
 import dispatcher from '../helpers/dispatcher.js';
-import globalUtils from '../helpers/globalutils.js';
+import globalUtils from '../helpers/utils/globalutils.js';
 import lazyRequest from '../helpers/lazyRequest.js';
 import session from '../helpers/session.js';
