+++ server/helpers/intents/permissions.js
