+++ server/helpers/session.js
@@ -1,10 +1,10 @@
 import { constants, deflateSync } from 'zlib';
 
 import dispatcher from './dispatcher.js';
-import globalUtils from './globalutils.js';
+import globalUtils from './utils/globalutils.js';
 import Intents from './intents.js';
 import lazyRequest from './lazyRequest.js';
-import { logText } from './logger.ts';
+import { logText } from './utils/logger.ts';
 
 let erlpack = null;
