+++ server/protocol/mrserver.ts
@@ -1,8 +1,8 @@
 import { WebSocketServer } from 'ws';
 
-import { mrHandlers, OPCODES } from './handlers/mr.ts';
-import { logText } from './helpers/logger.ts';
-import { type GatewayPayload, GatewayPayloadSchema } from './types/gateway.ts';
+import { mrHandlers, OPCODES } from '../handlers/mr.ts';
+import { logText } from '../helpers/utils/logger.ts';
+import { type GatewayPayload, GatewayPayloadSchema } from '../types/gateway.ts';
 
 // TODO: Replace all String() or "as type" conversions with better ones
