+++ server/protocol/rtcserver.ts
@@ -2,9 +2,9 @@ import sodium from 'libsodium-wrappers';
 import { EventEmitter } from 'node:events';
 import { WebSocketServer } from 'ws';
 
-import { OPCODES, rtcHandlers } from './handlers/rtc.ts';
-import { logText } from './helpers/logger.ts';
-import { type GatewayPayload, GatewayPayloadSchema } from './types/gateway.ts';
+import { OPCODES, rtcHandlers } from '../handlers/rtc.ts';
+import { logText } from '../helpers/utils/logger.ts';
+import { type GatewayPayload, GatewayPayloadSchema } from '../types/gateway.ts';
 
 // TODO: Replace all String() or "as type" conversions with better ones
